refactoring the file/class names as Ishikawa-san suggested
