+++ control/autoware_simple_pure_pursuit/src/simple_pure_pursuit.hpp
@@ -12,53 +12,86 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SIMPLE_PURE_PURSUIT_HPP_
-#define SIMPLE_PURE_PURSUIT_HPP_
-
-#include "simple_pure_pursuit_core_logic.hpp"
-
-#include <autoware_utils_rclcpp/polling_subscriber.hpp>
-#include <autoware_vehicle_info_utils/vehicle_info_utils.hpp>
-#include <rclcpp/rclcpp.hpp>
+#ifndef SIMPLE_PURE_PURSUIT_CORE_logic_HPP_
+#define SIMPLE_PURE_PURSUIT_CORE_logic_HPP_
 
 #include <autoware_control_msgs/msg/control.hpp>
 #include <autoware_planning_msgs/msg/trajectory.hpp>
-#include <autoware_planning_msgs/msg/trajectory_point.hpp>
 #include <nav_msgs/msg/odometry.hpp>
 
-#include <memory>
-
-namespace autoware::control::simple_pure_pursuit
+namespace autoware::control::simple_pure_pursuit_node
 {
-using autoware_planning_msgs::msg::Trajectory;
-using autoware_planning_msgs::msg::TrajectoryPoint;
-using nav_msgs::msg::Odometry;
 
-class SimplePurePursuitNode : public rclcpp::Node
+/**
+@ brief Params for SimplePurePursuitNode class
+*
+* @param lookahead_gain Gain for calculating lookahead distance
+* @param lookahead_min_distance Minimum lookahead distance
+* @param speed_proportional_gain Proportional gain for longitudinal control
+* @param use_external_target_vel Whether to use external target velocity or not
+* @param external_target_vel External target velocity to use when use_external_target_vel is true
+* @param wheel_base_m Wheel base of vehicle in meters
+*/
+struct SimplePurePursuitParameters
 {
-public:
-  explicit SimplePurePursuitNode(const rclcpp::NodeOptions & node_options);
-
-private:
-  // subscribers
-  autoware_utils_rclcpp::InterProcessPollingSubscriber<Odometry> odom_sub_{
-    this, "~/input/odometry"};
-  autoware_utils_rclcpp::InterProcessPollingSubscriber<Trajectory> traj_sub_{
-    this, "~/input/trajectory"};
-
-  // publishers
-  rclcpp::Publisher<autoware_control_msgs::msg::Control>::SharedPtr pub_control_command_;
-
-  // timer
-  rclcpp::TimerBase::SharedPtr timer_;
-
-  // simple_pure_pursuit core logic
-  std::unique_ptr<SimplePurePursuitCorelogic> core_logic_;
-
-  // functions
-  void on_timer();
+  double lookahead_gain;
+  double lookahead_min_distance;
+  double speed_proportional_gain;
+  bool use_external_target_vel;
+  double external_target_vel;
+  double wheel_base_m;
 };
 
-}  // namespace autoware::control::simple_pure_pursuit
+class SimplePurePursuit
+{
+public:
+  // Constructor
+  explicit SimplePurePursuit(const SimplePurePursuitParameters & params);
 
-#endif  // SIMPLE_PURE_PURSUIT_HPP_
+  // Set param struct for algorithm
+  void set_params(const SimplePurePursuitParameters & params);
+
+  /**
+   * @brief Create control command based on current odom and traj
+   *
+   * @param odom Current odometry of vehicle
+   * @param traj Current trajectory to follow
+   *
+   * @return Control command to be published
+   */
+  [[nodiscard]] autoware_control_msgs::msg::Control create_control_command(
+    const nav_msgs::msg::Odometry & odom,
+    const autoware_planning_msgs::msg::Trajectory & traj) const;
+
+private:
+  SimplePurePursuitParameters params_;
+
+  /**
+   * @brief Calculate longitudinal control command
+   *
+   * @param odom Current odometry of vehicle
+   * @param target_longitudinal_vel Target longitudinal velocity to achieve
+   *
+   * @return Longitudinal control command
+   */
+  [[nodiscard]] autoware_control_msgs::msg::Longitudinal calc_longitudinal_control(
+    const nav_msgs::msg::Odometry & odom, const double target_longitudinal_vel) const;
+
+  /**
+   * @brief Calculate lateral control command
+   *
+   * @param odom Current odometry of vehicle
+   * @param traj Current trajectory to follow
+   * @param target_longitudinal_vel Target longitudinal velocity to achieve
+   * @param closest_traj_point_idx Index of closest trajectory point to current vehicle position
+   *
+   * @return Lateral control command
+   */
+  [[nodiscard]] autoware_control_msgs::msg::Lateral calc_lateral_control(
+    const nav_msgs::msg::Odometry & odom, const autoware_planning_msgs::msg::Trajectory & traj,
+    const double target_longitudinal_vel, const size_t closest_traj_point_idx) const;
+};
+
+};  // namespace autoware::control::simple_pure_pursuit_node
+
+#endif  // SIMPLE_PURE_PURSUIT_CORE_logic_HPP_
